DescribedEntitiesListView renamed to NameSelectListView LocalTopicsListView renamed to EditRemoveListView
EditRemoveListView:
- added field 'service' with setter
- added method 'removeItem' which removes from both service and items list
