+++ src/main/java/egl/client/controller/topic/TopicTasksController.java
@@ -7,7 +7,7 @@ import egl.client.controller.Controller;
 import egl.client.controller.task.TaskController;
 import egl.client.model.topic.LocalTopic;
 import egl.client.service.FxmlService;
-import egl.client.view.table.DescribedEntitiesListView;
+import egl.client.view.table.NameSelectListView;
 import egl.core.model.task.Task;
 import javafx.fxml.FXML;
 import lombok.RequiredArgsConstructor;
@@ -21,7 +21,7 @@ public class TopicTasksController implements Controller {
 
     private final FxmlService fxmlService;
 
-    @FXML private DescribedEntitiesListView<Task> tasksListView;
+    @FXML private NameSelectListView<Task> tasksListView;
 
     private LocalTopic controllerTopic;
