+++ src/main/resources/egl/client/view/table/NameSelectListView.fxml
